refactor voting card
